requirements.txt
streamlit
requests
requests-oauthlib
google-auth
google-auth-oauthlib
pandas
groq
sqlalchemy